Made Economy interface use BigDecimal to avoid precision loss

+++ src/main/java/gvlfm78/plugin/hotels/bukkit/BukkitEconomyAdapter.java
@@ -5,18 +5,20 @@ import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
 
+import java.math.BigDecimal;
+
 public class BukkitEconomyAdapter implements EconomyAdapter {
 
     private static Economy economy = Bukkit.getServer().getServicesManager().getRegistration(Economy.class).getProvider();
 
     @Override
-    public double getBalance(OfflinePlayer player) {
-        return economy.getBalance(player);
+    public BigDecimal getBalance(OfflinePlayer player) {
+        return BigDecimal.valueOf(economy.getBalance(player));
     }
 
     @Override
-    public boolean withdrawAmount(OfflinePlayer player, double amount) {
-        return economy.withdrawPlayer(player, amount).transactionSuccess();
+    public boolean withdrawAmount(OfflinePlayer player, BigDecimal amount) {
+        return economy.withdrawPlayer(player, amount.doubleValue()).transactionSuccess();
     }
 
     @Override
@@ -30,7 +32,7 @@ public class BukkitEconomyAdapter implements EconomyAdapter {
     }
 
     @Override
-    public String formatCurrency(double amount){
-        return economy.format(amount);
+    public String formatCurrency(BigDecimal amount){
+        return economy.format(amount.doubleValue());
     }
 }
